Rename QtMainApplication directory
