+++ src/CMakeLists.txt
@@ -3,6 +3,6 @@ add_subdirectory(Spectrogram)
 
 add_subdirectory(CLIApplication)
 
-add_subdirectory(QtMainApplication)
+add_subdirectory(QtApplication)
 
 add_subdirectory(GtkmmApplication)
