+++ src/QtApplication/QtApplication.cpp
