+++ src/QtApplication/QtApplication.h
